Fix deployment

+++ db/migrate/20240107093531_update_ingredients_name_index.rb
@@ -1,6 +1,8 @@
 class UpdateIngredientsNameIndex < ActiveRecord::Migration[7.1]
   def change
-    remove_index :ingredients, :name
+    if index_exists?(:ingredients, :name)
+      remove_index :ingredients, :name
+    end
     add_index :ingredients, [ :name, :created_by_id ], unique: true
   end
 end
